+++ qa-support/src/main/java/gir/test/util/AbstractGirTest.java
@@ -3,6 +3,7 @@ package gir.test.util;
 import gir.GirContext;
 import gir.GirTestUtil;
 import gir.io.FileUtil;
+import java.nio.file.Path;
 import org.realityforge.braincheck.BrainCheckTestUtil;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
@@ -10,12 +11,16 @@ import org.testng.annotations.BeforeMethod;
 @SuppressWarnings( "SameParameterValue" )
 public abstract class AbstractGirTest
 {
+  private Path _workingDirectory;
+
   @BeforeMethod
   protected void beforeTest()
     throws Exception
   {
     BrainCheckTestUtil.resetConfig( false );
-    FileUtil.setCurrentDirectory( FileUtil.cwd() );
+    resetWorkingDirectory();
+    _workingDirectory = FileUtil.createTempDir();
+    FileUtil.setCurrentDirectory( _workingDirectory );
     GirTestUtil.setContext( new GirContext() );
   }
 
@@ -23,5 +28,16 @@ public abstract class AbstractGirTest
   protected void afterTest()
   {
     BrainCheckTestUtil.resetConfig( true );
+    resetWorkingDirectory();
+  }
+
+  private void resetWorkingDirectory()
+  {
+    if( null != _workingDirectory )
+    {
+      final Path directory = _workingDirectory;
+      _workingDirectory = null;
+      FileUtil.deleteDirIfExists( directory );
+    }
   }
 }