+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `FileUtil.write(...)` methods to ease writing files relative to current directory.
 * Add `FileUtil.createTempDir()` helper method.
 * Add `FileUtil.inTempDir(Action)` helper method.
+* Refactor tests to run in an isolated, per-test, temp directory.
 
 ### [v0.07](https://github.com/realityforge/gir/tree/v0.07) (2019-02-04)
 [Full Changelog](https://github.com/realityforge/gir/compare/v0.06...v0.07)